Added Map & adjusted collision detection
Collision detection needs to be adjusted a bit, seems to have a 1-off
error as the edges seem to overlap a bit.

Will need to implement some logic to aid in turning, e.g. if you are in
a cell and the user clicks into a direction before he actually gets to
the middle of that cell, then he should turn into that direction once he
gets there.